+++ examples/basic-tracer-node/multi_exporter.js
@@ -1,12 +1,12 @@
 'use strict';
 
 const opentelemetry = require('@opentelemetry/core');
-const { BasicTracerRegistry, BatchSpanProcessor, SimpleSpanProcessor } = require('@opentelemetry/tracing');
+const { BasicTracerProvider, BatchSpanProcessor, SimpleSpanProcessor } = require('@opentelemetry/tracing');
 const { JaegerExporter } = require('@opentelemetry/exporter-jaeger');
 const { ZipkinExporter } = require('@opentelemetry/exporter-zipkin');
 const { CollectorExporter } = require('@opentelemetry/exporter-collector');
 
-const registry = new BasicTracerRegistry();
+const provider = new BasicTracerProvider();
 
 const zipkinExporter = new ZipkinExporter({ serviceName: 'basic-service' });
 const jaegerExporter = new JaegerExporter({
@@ -16,7 +16,7 @@ const collectorExporter = new CollectorExporter({ serviceName: 'basic-service' }
 
 // It is recommended to use this BatchSpanProcessor for better performance
 // and optimization, especially in production.
-registry.addSpanProcessor(new BatchSpanProcessor(zipkinExporter, {
+provider.addSpanProcessor(new BatchSpanProcessor(zipkinExporter, {
   // This is added for example, default size is 100.
   bufferSize: 10,
 }));
@@ -27,10 +27,10 @@ tracer.addSpanProcessor(new BatchSpanProcessor(jaegerExporter), {
   bufferSize: 10,
 });
 
-registry.addSpanProcessor(new SimpleSpanProcessor(collectorExporter));
+provider.addSpanProcessor(new SimpleSpanProcessor(collectorExporter));
 
-// Initialize the OpenTelemetry APIs to use the BasicTracerRegistry bindings
-opentelemetry.initGlobalTracerRegistry(registry);
+// Initialize the OpenTelemetry APIs to use the BasicTracerProvider bindings
+opentelemetry.initGlobalTracerProvider(provider);
 
 // Create a span. A span must be closed.
 const parentSpan = tracer.startSpan('main');