+++ examples/opentracing-shim/shim.js
@@ -1,17 +1,17 @@
 'use strict';
 
-const { NodeTracerRegistry } = require('@opentelemetry/node');
+const { NodeTracerProvider } = require('@opentelemetry/node');
 const { SimpleSpanProcessor } = require('@opentelemetry/tracing');
 const { JaegerExporter } = require('@opentelemetry/exporter-jaeger');
 const { ZipkinExporter } = require('@opentelemetry/exporter-zipkin');
 const { TracerShim } = require('@opentelemetry/shim-opentracing');
 
 function shim(serviceName) {
-  const registry = new NodeTracerRegistry();
+  const provider = new NodeTracerProvider();
 
-  registry.addSpanProcessor(new SimpleSpanProcessor(getExporter(serviceName)));
+  provider.addSpanProcessor(new SimpleSpanProcessor(getExporter(serviceName)));
 
-  return new TracerShim(registry.getTracer('opentracing-shim'));
+  return new TracerShim(provider.getTracer('opentracing-shim'));
 }
 
 function getExporter(serviceName) {