+++ packages/opentelemetry-plugin-dns/test/integrations/dnspromise-lookup.test.ts
@@ -20,7 +20,7 @@ import {
 } from '@opentelemetry/tracing';
 import * as assert from 'assert';
 import { NoopLogger } from '@opentelemetry/core';
-import { NodeTracerRegistry } from '@opentelemetry/node';
+import { NodeTracerProvider } from '@opentelemetry/node';
 import { plugin } from '../../src/dns';
 import * as dns from 'dns';
 import * as utils from '../utils/utils';
@@ -30,8 +30,8 @@ import { CanonicalCode } from '@opentelemetry/api';
 
 const memoryExporter = new InMemorySpanExporter();
 const logger = new NoopLogger();
-const registry = new NodeTracerRegistry({ logger });
-registry.addSpanProcessor(new SimpleSpanProcessor(memoryExporter));
+const provider = new NodeTracerProvider({ logger });
+provider.addSpanProcessor(new SimpleSpanProcessor(memoryExporter));
 
 describe('dns.promises.lookup()', () => {
   before(function(done) {
@@ -42,7 +42,7 @@ describe('dns.promises.lookup()', () => {
 
     // if node version is supported, it's mandatory for CI
     if (process.env.CI) {
-      plugin.enable(dns, registry, registry.logger);
+      plugin.enable(dns, provider, provider.logger);
       done();
       return;
     }
@@ -54,7 +54,7 @@ describe('dns.promises.lookup()', () => {
       }
       done();
     });
-    plugin.enable(dns, registry, registry.logger);
+    plugin.enable(dns, provider, provider.logger);
   });
 
   afterEach(() => {