+++ packages/opentelemetry-plugin-http/test/functionals/http-disable.test.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { NoopTracerRegistry, NOOP_TRACER } from '@opentelemetry/api';
+import { NoopTracerProvider, NOOP_TRACER } from '@opentelemetry/api';
 import { NoopLogger } from '@opentelemetry/core';
 import * as assert from 'assert';
 import * as http from 'http';
@@ -30,12 +30,12 @@ describe('HttpPlugin', () => {
 
   describe('disable()', () => {
     const logger = new NoopLogger();
-    const registry = new NoopTracerRegistry();
+    const provider = new NoopTracerProvider();
     before(() => {
       nock.cleanAll();
       nock.enableNetConnect();
 
-      plugin.enable(http, registry, logger);
+      plugin.enable(http, provider, logger);
       // Ensure that http module is patched.
       assert.strictEqual(http.Server.prototype.emit.__wrapped, true);
       server = http.createServer((request, response) => {
@@ -61,7 +61,7 @@ describe('HttpPlugin', () => {
       server.close();
     });
     describe('unpatch()', () => {
-      it('should not call registry methods for creating span', async () => {
+      it('should not call provider methods for creating span', async () => {
         plugin.disable();
         const testPath = '/incoming/unpatch/';